+++ bjc_springboot/src/main/java/com/lambdacode/spring/boot/crud/repository/ReservaRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReservaRepository extends JpaRepository<Reserva, Long> {
+public interface ReservaRepository extends JpaRepository<Reserva, Integer> {
     Optional<Reserva> findByUsuario_IdUsuarioAndStatus(Integer idUsuario, StatusBicicleta status);
     boolean existsByUsuario_IdUsuarioAndStatusIn(Integer idUsuario, List<StatusBicicleta> status);
 }
